+++ packages/browser/src/login/oidc/redirectHandler/AuthCodeRedirectHandler.ts
@@ -60,62 +60,6 @@ async function verifyIdToken(
   return validateIdToken(idToken, jwks, issuer, client.clientId);
 }
 
-// A lifespan of 30 minutes is ESS's default. This could be removed if we
-// configure the server to return the remaining lifespan of the cookie.
-export const DEFAULT_LIFESPAN = 30 * 60 * 1000;
-
-/**
- * Stores the resource server session information in local storage, so that they
- * can be used on refresh.
- * @param webId
- * @param authenticatedFetch
- * @param storageUtility
- */
-async function setupResourceServerSession(
-  webId: string,
-  authenticatedFetch: typeof fetch,
-  storageUtility: IStorageUtility
-): Promise<void> {
-  const webIdAsUrl = new URL(webId);
-  const resourceServerIri = webIdAsUrl.origin;
-  // Querying the /session endpoint does not set the cookie, but issuing an
-  // authenticated request to any actual resource (even public ones) does,
-  // so we fetch the user's WebID before checking the /session endpoint.
-  await authenticatedFetch(webId);
-  try {
-    const resourceServerResponse = await authenticatedFetch(
-      `${resourceServerIri}/session`
-    );
-
-    if (resourceServerResponse.status === 200) {
-      await storageUtility.storeResourceServerSessionInfo(
-        webId,
-        resourceServerIri,
-        // Note that here, if the lifespan of the cookie was returned by the
-        // server, we'd expect a relative value (the remaining time of validity)
-        // rather than an absolute one (the moment when the cookie expires).
-        Date.now() + DEFAULT_LIFESPAN
-      );
-      return;
-    }
-    // In this case, the resource server either:
-    // - does not have the expected endpoint, or
-    // - does not recognize the user
-    // Either way, no cookie is expected to be set there, and any existing
-    // session information should be cleared.
-    await storageUtility.clearResourceServerSessionInfo(resourceServerIri);
-  } catch (_e) {
-    // Setting the `credentials=include` option on fetch, which is required in
-    // the current approach based on a RS cookie, may result in an error if
-    // attempting to access an URL, depending on the CORS policies.
-    // Since this internal fetch is necessary, and out of control of the
-    // calling library, there is no other solution but to swallow the exception.
-    // This may happen depending on how the target RS handles a request to the
-    // /session endpoint.
-    await storageUtility.clearResourceServerSessionInfo(resourceServerIri);
-  }
-}
-
 /**
  * @hidden
  */
@@ -254,18 +198,6 @@ export class AuthCodeRedirectHandler implements IRedirectHandler {
         secure: false,
       }
     );
-    // TODO: This is a temporary workaround. When deprecating the cookie-based auth,
-    // this should be all cleared.
-    const essWorkaroundDisabled =
-      window.localStorage.getItem("tmp-resource-server-session-enabled") ===
-      "false";
-    if (!essWorkaroundDisabled) {
-      await setupResourceServerSession(
-        tokens.webId,
-        authFetch,
-        this.storageUtility
-      );
-    }
 
     const sessionInfo = await this.sessionInfoManager.get(storedSessionId);
     if (!sessionInfo) {