+++ packages/core/src/index.ts
@@ -70,7 +70,6 @@ export {
   loadOidcContextFromStorage,
   saveSessionInfoToStorage,
   getSessionIdFromOauthState,
-  ResourceServerSession,
 } from "./storage/StorageUtility";
 export { default as InMemoryStorage } from "./storage/InMemoryStorage";